Revert "INK-215: Return invalid produce results if no other batch succeeds"
This reverts commit 5b3254d6c0b60726f4d5b66cf5c747903f1273ee.
