+++ storage/inkless/src/main/java/io/aiven/inkless/produce/FileCommitter.java
@@ -17,7 +17,6 @@
  */
 package io.aiven.inkless.produce;
 
-import io.aiven.inkless.control_plane.CommitBatchResponse;
 import org.apache.kafka.common.utils.Time;
 
 import com.groupcdg.pitest.annotations.DoNotMutate;
@@ -26,8 +25,6 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Collections;
-import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
@@ -67,7 +64,6 @@ class FileCommitter implements Closeable {
     private final Duration fileUploadRetryBackoff;
     private final ExecutorService executorServiceUpload;
     private final ExecutorService executorServiceCommit;
-    private final ExecutorService executorServiceComplete;
     private final ExecutorService executorServiceCacheStore;
 
     private final AtomicInteger totalFilesInProgress = new AtomicInteger(0);
@@ -86,8 +82,7 @@ class FileCommitter implements Closeable {
         this(brokerId, controlPlane, objectKeyCreator, storage, keyAlignmentStrategy, objectCache, time, maxFileUploadAttempts, fileUploadRetryBackoff,
             Executors.newCachedThreadPool(new InklessThreadFactory("inkless-file-uploader-", false)),
             // It must be single-thread to preserve the commit order.
-            Executors.newSingleThreadExecutor(new InklessThreadFactory("inkless-file-committer-", false)),
-            Executors.newCachedThreadPool(new InklessThreadFactory("inkless-file-uploader-finisher-", false)),
+            Executors.newSingleThreadExecutor(new InklessThreadFactory("inkless-file-uploader-finisher-", false)),
             Executors.newCachedThreadPool(new InklessThreadFactory("inkless-file-cache-store-", false)),
             new FileCommitterMetrics(time)
         );
@@ -105,7 +100,6 @@ class FileCommitter implements Closeable {
                   final Duration fileUploadRetryBackoff,
                   final ExecutorService executorServiceUpload,
                   final ExecutorService executorServiceCommit,
-                  final ExecutorService executorServiceComplete,
                   final ExecutorService executorServiceCacheStore,
                   final FileCommitterMetrics metrics) {
         this.brokerId = brokerId;
@@ -125,8 +119,6 @@ class FileCommitter implements Closeable {
             "executorServiceUpload cannot be null");
         this.executorServiceCommit = Objects.requireNonNull(executorServiceCommit,
             "executorServiceCommit cannot be null");
-        this.executorServiceComplete = Objects.requireNonNull(executorServiceComplete,
-                "executorServiceComplete cannot be null");
         this.executorServiceCacheStore = Objects.requireNonNull(executorServiceCacheStore,
                 "executorServiceCacheStore cannot be null");
 
@@ -138,62 +130,53 @@ class FileCommitter implements Closeable {
 
     void commit(final ClosedFile file) throws InterruptedException {
         Objects.requireNonNull(file, "file cannot be null");
+        metrics.fileAdded(file.size());
 
         lock.lock();
         try {
             final Instant uploadAndCommitStart = TimeUtils.durationMeasurementNow(time);
-            Future<List<CommitBatchResponse>> commitFuture;
-            if (file.isEmpty()) {
-                // If the file is empty, skip uploading and committing, but proceed with the later steps.
-                commitFuture = CompletableFuture.completedFuture(Collections.emptyList());
-            } else {
-                metrics.fileAdded(file.size());
-                totalFilesInProgress.addAndGet(1);
-                totalBytesInProgress.addAndGet(file.size());
 
-                // Start uploading and add to the commit queue (as Runnable).
-                // This ensures files are uploaded in concurrently, but committed to the control plane sequentially,
-                // because `executorServiceCommit` is single-threaded.
-                final FileUploadJob uploadJob = FileUploadJob.createFromByteArray(
-                        objectKeyCreator,
-                        storage,
-                        time,
-                        maxFileUploadAttempts,
-                        fileUploadRetryBackoff,
-                        file.data(),
-                        metrics::fileUploadFinished
-                );
-                final Future<ObjectKey> uploadFuture = executorServiceUpload.submit(uploadJob);
+            totalFilesInProgress.addAndGet(1);
+            totalBytesInProgress.addAndGet(file.size());
 
-                final FileCommitJob commitJob = new FileCommitJob(
-                        brokerId,
-                        file,
-                        uploadFuture,
-                        time,
-                        controlPlane,
-                        storage,
-                        metrics::fileCommitFinished
-                );
-                commitFuture = executorServiceCommit.submit(commitJob);
+            // Start uploading and add to the commit queue (as Runnable).
+            // This ensures files are uploaded in concurrently, but committed to the control plane sequentially,
+            // because `executorServiceCommit` is single-threaded.
+            final FileUploadJob uploadJob = FileUploadJob.createFromByteArray(
+                objectKeyCreator,
+                storage,
+                time,
+                maxFileUploadAttempts,
+                fileUploadRetryBackoff,
+                file.data(),
+                metrics::fileUploadFinished
+            );
+            final Future<ObjectKey> uploadFuture = executorServiceUpload.submit(uploadJob);
 
-
-                final CacheStoreJob cacheStoreJob = new CacheStoreJob(
-                        time,
-                        objectCache,
-                        keyAlignmentStrategy,
-                        file.data(),
-                        uploadFuture,
-                        metrics::cacheStoreFinished
-                );
-                CompletableFuture.runAsync(cacheStoreJob, executorServiceCacheStore)
-                    .whenComplete((result, error) -> {
-                        totalFilesInProgress.addAndGet(-1);
-                        totalBytesInProgress.addAndGet(-file.size());
-                        metrics.fileFinished(file.start(), uploadAndCommitStart);
-                    });
-            }
-            final AppendCompleterJob completerJob = new AppendCompleterJob(file, commitFuture, time, metrics::appendCompletionFinished);
-            executorServiceComplete.submit(completerJob);
+            final FileCommitJob commitJob = new FileCommitJob(
+                brokerId,
+                file,
+                uploadFuture,
+                time,
+                controlPlane,
+                storage,
+                metrics::fileCommitFinished
+            );
+            CompletableFuture.runAsync(commitJob, executorServiceCommit)
+                .whenComplete((result, error) -> {
+                    totalFilesInProgress.addAndGet(-1);
+                    totalBytesInProgress.addAndGet(-file.size());
+                    metrics.fileFinished(file.start(), uploadAndCommitStart);
+                });
+            final CacheStoreJob cacheStoreJob = new CacheStoreJob(
+                time,
+                objectCache,
+                keyAlignmentStrategy,
+                file.data(),
+                uploadFuture,
+                metrics::cacheStoreFinished
+            );
+            executorServiceCacheStore.submit(cacheStoreJob);
         } finally {
             lock.unlock();
         }
