+++ storage/inkless/src/main/java/io/aiven/inkless/produce/FileCommitterMetrics.java
@@ -93,10 +93,6 @@ class FileCommitterMetrics implements Closeable {
         fileCommitRate.increment();
     }
 
-    public void appendCompletionFinished(final long durationMs) {
-
-    }
-
     void fileFinished(final Instant fileStart, final Instant uploadAndCommitStart) {
         final Instant now = TimeUtils.durationMeasurementNow(time);
         fileTotalLifeTimeHistogram.update(Duration.between(fileStart, now).toMillis());
